+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog for autoNumeric
 
+### 4.2.9
++ Fixes #568 Using brackets for negative numbers in AutoNumeric.format returns "undefined"
+
+### 4.2.8
++ Fixes #566 Add the Turkish predefined currency
+
 ### 4.2.7
 + Fixes #521 The `input` event is not fired on `paste` if the element is empty or is completely selected beforehand
 + Fixes #563 The `import AutoNumeric from 'AutoNumeric'` line in the readme does not work on case sensitive OS (like Linux)